feat(graphs): added a new graph type
With this new type it is posible to create and manipulate graphs
A basic Trace function is added to find a maximal path in an undirected
graph. In this case, maximal does not implies longest, only that the
path cannot be extended because all incident edges to the last vertex in
the path have been used.